+++ src/wetwire_gitlab/testing/__init__.py
@@ -1,0 +1,7 @@
+"""Testing utilities for wetwire-gitlab."""
+
+from .semantic_compare import compare_yaml_semantic
+
+__all__ = [
+    "compare_yaml_semantic",
+]